backend/image_analysis.py
"""TerraQuant AI - Image Analysis Service

Supports both satellite imagery and regular photos for carbon-oriented insights.
"""

import random
from datetime import datetime
from pathlib import Path

from PIL import Image, ImageStat


def analyze_satellite_image(image_path: str):
    """Analyze uploaded image and return carbon-related insights.

    The API name is kept for backward compatibility, but behavior now supports:
    - Satellite images (land-cover oriented)
    - Normal photos (scene/emission oriented)
    """
    image_name = Path(image_path).name.lower()
    image_mode = _detect_image_mode(image_name)
    image_signals = _extract_image_signals(image_path)

    if image_mode == "satellite":
        return _analyze_satellite_mode(image_path, image_name, image_signals)
    return _analyze_photo_mode(image_path, image_name, image_signals)


def _detect_image_mode(image_name: str) -> str:
    satellite_tokens = [
        "satellite", "landsat", "sentinel", "eo", "aerial", "drone",
        "map", "terrain", "ndvi", "geospatial"
    ]
    if any(token in image_name for token in satellite_tokens):
        return "satellite"
    return "photo"


def _extract_image_signals(image_path: str):
    try:
        with Image.open(image_path) as img:
            rgb = img.convert("RGB")
            stat = ImageStat.Stat(rgb)
            mean_rgb = stat.mean
            brightness = sum(mean_rgb) / 3
            pixel_count = rgb.size[0] * rgb.size[1]
            return {
                "width": rgb.size[0],
                "height": rgb.size[1],
                "pixels": pixel_count,
                "brightness": brightness,
                "mean_r": mean_rgb[0],
                "mean_g": mean_rgb[1],
                "mean_b": mean_rgb[2],
            }
    except Exception:
        # Fallback signals when image metadata parsing fails.
        return {
            "width": 1280,
            "height": 720,
            "pixels": 1280 * 720,
            "brightness": 120,
            "mean_r": 120,
            "mean_g": 125,
            "mean_b": 115,
        }


def _analyze_satellite_mode(image_path: str, image_name: str, image_signals: dict):
    forest_pct = round(random.uniform(25, 55), 1)
    agriculture_pct = round(random.uniform(15, 35), 1)
    urban_pct = round(random.uniform(10, 25), 1)
    water_pct = round(random.uniform(3, 15), 1)
    barren_pct = round(random.uniform(2, 10), 1)

    if "forest" in image_name or "tree" in image_name or "amazon" in image_name:
        forest_pct = round(random.uniform(60, 85), 1)
        urban_pct = round(random.uniform(2, 8), 1)
        agriculture_pct = round(random.uniform(5, 15), 1)
    elif "farm" in image_name or "crop" in image_name or "agriculture" in image_name:
        agriculture_pct = round(random.uniform(55, 80), 1)
        forest_pct = round(random.uniform(5, 20), 1)
        urban_pct = round(random.uniform(3, 10), 1)
    elif "city" in image_name or "urban" in image_name or "building" in image_name:
        urban_pct = round(random.uniform(45, 75), 1)
        forest_pct = round(random.uniform(5, 15), 1)
        agriculture_pct = round(random.uniform(3, 12), 1)
    elif "desert" in image_name or "barren" in image_name or "dry" in image_name:
        barren_pct = round(random.uniform(50, 80), 1)
        forest_pct = round(random.uniform(2, 10), 1)
        agriculture_pct = round(random.uniform(3, 12), 1)
    elif "coast" in image_name or "ocean" in image_name or "water" in image_name:
        water_pct = round(random.uniform(40, 70), 1)
        forest_pct = round(random.uniform(10, 25), 1)

    total = forest_pct + agriculture_pct + urban_pct + water_pct + barren_pct
    forest_pct = round((forest_pct / total) * 100, 1)
    agriculture_pct = round((agriculture_pct / total) * 100, 1)
    urban_pct = round((urban_pct / total) * 100, 1)
    water_pct = round((water_pct / total) * 100, 1)
    barren_pct = round((barren_pct / total) * 100, 1)

    carbon_density = (
        forest_pct * 1.8
        + agriculture_pct * 0.8
        + urban_pct * 0.2
        + water_pct * 0.1
        + barren_pct * 0.1
    )
    vegetation_health = (
        forest_pct * 0.9
        + agriculture_pct * 0.7
        + urban_pct * 0.2
        + water_pct * 0.1
        + barren_pct * 0.1
    )

    return {
        "analysis_mode": "satellite",
        "scene_type": "terrain-land-cover",
        "land_cover": {
            "forest": forest_pct,
            "agriculture": agriculture_pct,
            "urban": urban_pct,
            "water": water_pct,
            "barren": barren_pct,
        },
        "carbon_density": round(carbon_density, 1),
        "vegetation_health": round(min(vegetation_health, 100), 1),
        "estimated_daily_emission_kg": round(max(80 - carbon_density, 5) * 2.5, 1),
        "carbon_impact_score": round(max(100 - carbon_density / 2.2, 20), 1),
        "emission_hotspots": _infer_emission_hotspots(image_name, "satellite", image_signals),
        "analysis_date": datetime.now().strftime("%Y-%m-%d"),
        "image_analyzed": image_path,
        "recommendations": _get_satellite_recommendations(
            forest_pct,
            agriculture_pct,
            urban_pct,
            carbon_density,
        ),
    }


def _analyze_photo_mode(image_path: str, image_name: str, image_signals: dict):
    brightness = image_signals["brightness"]
    greenness_delta = image_signals["mean_g"] - ((image_signals["mean_r"] + image_signals["mean_b"]) / 2)

    scene_type = "mixed-activity"
    urban_pct = 35.0
    forest_pct = 22.0
    agriculture_pct = 18.0
    water_pct = 10.0
    barren_pct = 15.0

    if any(token in image_name for token in ["traffic", "car", "road", "factory", "plant", "smoke"]):
        scene_type = "urban-industrial"
        urban_pct, forest_pct, agriculture_pct, water_pct, barren_pct = 62.0, 10.0, 8.0, 5.0, 15.0
    elif any(token in image_name for token in ["park", "tree", "forest", "garden", "farm"]):
        scene_type = "green-landscape"
        urban_pct, forest_pct, agriculture_pct, water_pct, barren_pct = 16.0, 50.0, 24.0, 5.0, 5.0
    elif any(token in image_name for token in ["sea", "coast", "river", "lake", "ocean"]):
        scene_type = "waterfront"
        urban_pct, forest_pct, agriculture_pct, water_pct, barren_pct = 18.0, 18.0, 10.0, 46.0, 8.0

    if brightness < 90:
        urban_pct += 6.0
        forest_pct -= 2.0
        barren_pct += 2.0

    if greenness_delta > 8:
        forest_pct += 10.0
        urban_pct -= 6.0
        barren_pct -= 4.0

    total = forest_pct + agriculture_pct + urban_pct + water_pct + barren_pct
    forest_pct = round((forest_pct / total) * 100, 1)
    agriculture_pct = round((agriculture_pct / total) * 100, 1)
    urban_pct = round((urban_pct / total) * 100, 1)
    water_pct = round((water_pct / total) * 100, 1)
    barren_pct = round((barren_pct / total) * 100, 1)

    carbon_density = (
        forest_pct * 1.4
        + agriculture_pct * 0.7
        + urban_pct * 0.35
        + water_pct * 0.12
        + barren_pct * 0.18
    )
    vegetation_health = (
        forest_pct * 0.85
        + agriculture_pct * 0.65
        + urban_pct * 0.25
        + water_pct * 0.1
        + barren_pct * 0.08
    )

    estimated_daily_emission_kg = round(max(urban_pct * 2.1 - forest_pct * 0.7, 8), 1)
    carbon_impact_score = round(min(max(estimated_daily_emission_kg * 1.15, 10), 100), 1)

    return {
        "analysis_mode": "photo",
        "scene_type": scene_type,
        "land_cover": {
            "forest": forest_pct,
            "agriculture": agriculture_pct,
            "urban": urban_pct,
            "water": water_pct,
            "barren": barren_pct,
        },
        "carbon_density": round(carbon_density, 1),
        "vegetation_health": round(min(vegetation_health, 100), 1),
        "estimated_daily_emission_kg": estimated_daily_emission_kg,
        "carbon_impact_score": carbon_impact_score,
        "emission_hotspots": _infer_emission_hotspots(image_name, "photo", image_signals),
        "analysis_date": datetime.now().strftime("%Y-%m-%d"),
        "image_analyzed": image_path,
        "recommendations": _get_photo_recommendations(
            scene_type,
            carbon_impact_score,
            estimated_daily_emission_kg,
            forest_pct,
        ),
    }


def _infer_emission_hotspots(image_name: str, image_mode: str, image_signals: dict):
    hotspots = []
    if any(token in image_name for token in ["traffic", "car", "highway", "road"]):
        hotspots.append("Transport corridor congestion")
    if any(token in image_name for token in ["factory", "plant", "industrial", "smoke"]):
        hotspots.append("Industrial stack or process emissions")
    if any(token in image_name for token in ["waste", "landfill", "dump"]):
        hotspots.append("Waste handling and methane risk")
    if image_mode == "satellite" and image_signals["mean_g"] < image_signals["mean_r"]:
        hotspots.append("Low vegetation cluster detected")
    if image_signals["brightness"] < 85:
        hotspots.append("Potential heavy activity / low-light pollution zone")

    if not hotspots:
        hotspots = ["No major hotspot inferred from current image"]
    return hotspots


def _get_satellite_recommendations(forest_pct, agriculture_pct, urban_pct, carbon_density):
    recommendations = []

    if forest_pct > 50:
        recommendations.append("High forest coverage detected; prioritize conservation and anti-deforestation controls")
    elif forest_pct < 20:
        recommendations.append("Low forest coverage; plan reforestation and green-belt expansion")

    if agriculture_pct > 50:
        recommendations.append("Large agricultural footprint; adopt precision irrigation and regenerative practices")

    if urban_pct > 40:
        recommendations.append("High urban density; increase green roofs and district energy efficiency")

    if carbon_density < 50:
        recommendations.append("Low carbon density; prioritize sequestration projects and biodiversity restoration")
    elif carbon_density > 100:
        recommendations.append("Strong carbon sink profile; maintain and monitor for degradation")

    return recommendations if recommendations else ["Continue periodic satellite monitoring for land-use drift"]


def _get_photo_recommendations(scene_type, carbon_impact_score, estimated_daily_emission_kg, forest_pct):
    recommendations = []

    if scene_type == "urban-industrial":
        recommendations.append("Likely urban/industrial emissions; improve fuel efficiency and electrify operations")
        recommendations.append("Deploy stack monitoring and cleaner process controls")
    elif scene_type == "green-landscape":
        recommendations.append("Green coverage is positive; preserve vegetation and avoid land-use degradation")
    elif scene_type == "waterfront":
        recommendations.append("Protect waterfront ecosystems and monitor coastal resilience factors")

    if carbon_impact_score > 70:
        recommendations.append("High carbon impact score; prioritize immediate reduction interventions")

    if estimated_daily_emission_kg > 40:
        recommendations.append("Estimated daily emissions are elevated; assess logistics and equipment sources")

    if forest_pct < 20:
        recommendations.append("Add green buffers around high-activity zones to improve carbon balance")

    return recommendations if recommendations else ["Maintain periodic photo audits to track local emission patterns"]
