backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
groq>=1.1.0
Pillow>=10.0.0
